+++ sdk/storage/azure-storage-queues/test/ut/queue_client_test.cpp
@@ -247,14 +247,22 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_NO_THROW(queueClient.GetProperties());
 
     // default audience
-    clientOptions.Audience = Queues::Models::QueueAudience::DefaultAudience;
+    clientOptions.Audience = Queues::QueueAudience::DefaultAudience;
+    queueClient = Queues::QueueClient(m_queueClient->GetUrl(), credential, clientOptions);
+    EXPECT_NO_THROW(queueClient.GetProperties());
+
+    // service audience
+    auto keyCredential
+        = _internal::ParseConnectionString(StandardStorageConnectionString()).KeyCredential;
+    auto accountName = keyCredential->AccountName;
+    clientOptions.Audience = Queues::QueueAudience::CreateQueueServiceAccountAudience(accountName);
     queueClient = Queues::QueueClient(m_queueClient->GetUrl(), credential, clientOptions);
     EXPECT_NO_THROW(queueClient.GetProperties());
 
     // custom audience
     auto queueUrl = Azure::Core::Url(queueClient.GetUrl());
     clientOptions.Audience
-        = Queues::Models::QueueAudience(queueUrl.GetScheme() + "://" + queueUrl.GetHost());
+        = Queues::QueueAudience(queueUrl.GetScheme() + "://" + queueUrl.GetHost());
     queueClient = Queues::QueueClient(m_queueClient->GetUrl(), credential, clientOptions);
     EXPECT_NO_THROW(queueClient.GetProperties());
 
@@ -264,7 +272,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_NO_THROW(queueClient.GetProperties());
 
     // error audience
-    clientOptions.Audience = Queues::Models::QueueAudience("https://disk.compute.azure.com");
+    clientOptions.Audience = Queues::QueueAudience("https://disk.compute.azure.com");
     queueClient = Queues::QueueClient(m_queueClient->GetUrl(), credential, clientOptions);
     EXPECT_THROW(queueClient.GetProperties(), StorageException);