+++ sdk/storage/azure-storage-queues/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.2.0-beta.1 (Unreleased)
+## 12.2.0-beta.2 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,13 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 12.2.0-beta.1 (2023-10-17)
+
+### Features Added
+
+- Added new extendable enum `QueueAudience`
+- Added new field `Audience` in `QueueClientOptions`
 
 ## 12.1.0 (2023-09-12)